extract VVC boxes processing into its own file
