+++ libheif/box.cc
@@ -27,7 +27,7 @@
 #include "codecs/jpeg2000_boxes.h"
 #include "codecs/hevc_boxes.h"
 #include "codecs/mask_image.h"
-#include "codecs/vvc.h"
+#include "codecs/vvc_boxes.h"
 #include "codecs/avc_boxes.h"
 #include "codecs/avif_boxes.h"
 #include "codecs/tild.h"
